feat(mobile): split mobile into standalone @vesta/mobile workspace
Separates iOS/Android Tauri builds from the desktop workspace so each is
self-contained with no shared Rust. Mobile gets its own crate
(`vesta-mobile`, bundle id `com.vestarun.mobile`), CI wires the existing
iOS/Android jobs to `apps/mobile`, and `release.yml` bumps + refreshes the
new workspace's versions and lockfile.
