+++ .github/dependabot.yml
@@ -32,6 +32,14 @@ updates:
       tauri-deps:
         patterns: ["*"]
 
+  - package-ecosystem: cargo
+    directory: /apps/mobile/src-tauri
+    schedule:
+      interval: weekly
+    groups:
+      mobile-tauri-deps:
+        patterns: ["*"]
+
   - package-ecosystem: npm
     directory: /apps
     schedule:
