+++ .github/workflows/release.yml
@@ -74,16 +74,21 @@ jobs:
           sed -i "0,/^version = \"${CURRENT}\"/s|^version = \"${CURRENT}\"|version = \"${NEW}\"|" vestad/tests-integration/Cargo.toml
           sed -i "0,/^version = \"${CURRENT}\"/s|^version = \"${CURRENT}\"|version = \"${NEW}\"|" cli/Cargo.toml
           sed -i "0,/^version = \"${CURRENT}\"/s|^version = \"${CURRENT}\"|version = \"${NEW}\"|" apps/desktop/src-tauri/Cargo.toml
+          sed -i "0,/^version = \"${CURRENT}\"/s|^version = \"${CURRENT}\"|version = \"${NEW}\"|" apps/mobile/src-tauri/Cargo.toml
           sed -i "s|\"version\": \"${CURRENT}\"|\"version\": \"${NEW}\"|" apps/desktop/src-tauri/tauri.conf.json
+          sed -i "s|\"version\": \"${CURRENT}\"|\"version\": \"${NEW}\"|" apps/mobile/src-tauri/tauri.conf.json
           sed -i "s|\"version\": \"${CURRENT}\"|\"version\": \"${NEW}\"|" apps/web/package.json
           sed -i "s|\"version\": \"${CURRENT}\"|\"version\": \"${NEW}\"|" apps/desktop/package.json
+          sed -i "s|\"version\": \"${CURRENT}\"|\"version\": \"${NEW}\"|" apps/mobile/package.json
 
       - name: Refresh lockfiles
         run: |
-          (cd vestad && cargo check --quiet)
-          (cd cli && cargo check --quiet)
-          (cd apps/desktop/src-tauri && cargo check --quiet)
-          (cd agent && uv lock --quiet)
+          (cd vestad && cargo check --quiet) &
+          (cd cli && cargo check --quiet) &
+          (cd apps/desktop/src-tauri && cargo check --quiet) &
+          (cd apps/mobile/src-tauri && cargo check --quiet) &
+          (cd agent && uv lock --quiet) &
+          wait
 
       - name: Commit and push
         env:
