+++ apps/desktop/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@vesta/desktop",
   "private": true,
-  "version": "0.1.140",
+  "version": "0.1.144",
   "description": "Vesta desktop app (Tauri wrapper around @vesta/web)",
   "license": "MIT",
   "repository": {
