+++ apps/desktop/src-tauri/Cargo.toml
@@ -40,6 +40,3 @@ objc2-foundation = "0.3"
 
 [target.'cfg(target_os = "windows")'.dependencies]
 window-vibrancy = "0.7"
-
-[target.'cfg(target_os = "ios")'.dependencies]
-objc2 = "0.6"
