+++ apps/desktop/src-tauri/tauri.conf.json
@@ -65,9 +65,6 @@
     "resources": [
       "resources/*"
     ],
-    "iOS": {
-      "developmentTeam": "QB32XS5CT9"
-    },
     "macOS": {
       "entitlements": null
     }
