+++ apps/mobile/src-tauri/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "vesta-mobile"
+version = "0.1.144"
+description = "Vesta mobile app"
+license = "MIT"
+repository = "https://github.com/elyxlz/vesta"
+authors = ["elyxlz"]
+edition = "2021"
+
+[lib]
+name = "vesta_mobile"
+crate-type = ["lib", "cdylib", "staticlib"]
+
+[build-dependencies]
+tauri-build = { version = "2", features = [] }
+
+[dependencies]
+tauri = { version = "2", features = [] }
+serde_json = "1"
+tauri-plugin-dialog = "2.7.0"
+tauri-plugin-opener = "2.5.3"
+tauri-plugin-store = "2"
+tauri-plugin-notification = "2"
+
+[target.'cfg(target_os = "ios")'.dependencies]
+objc2 = "0.6"
+
+[profile.release]
+strip = true
+lto = "thin"
+codegen-units = 1
+opt-level = "s"
