+++ apps/mobile/src-tauri/capabilities/default.json
@@ -1,0 +1,16 @@
+{
+  "identifier": "default",
+  "description": "Default capability set for Vesta mobile",
+  "windows": ["main"],
+  "permissions": [
+    "core:default",
+    "dialog:allow-save",
+    "dialog:allow-open",
+    "opener:default",
+    "store:default",
+    "notification:default",
+    "notification:allow-is-permission-granted",
+    "notification:allow-request-permission",
+    "notification:allow-notify"
+  ]
+}
