+++ apps/mobile/src-tauri/icons/android/mipmap-anydpi-v26/ic_launcher.xml
