+++ apps/mobile/src-tauri/icons/android/mipmap-hdpi/ic_launcher_foreground.png
