+++ apps/mobile/src-tauri/icons/android/mipmap-hdpi/ic_launcher_round.png
