+++ apps/mobile/src-tauri/icons/android/mipmap-mdpi/ic_launcher.png
