+++ apps/mobile/src-tauri/icons/android/mipmap-xhdpi/ic_launcher_round.png
