+++ apps/mobile/src-tauri/icons/android/mipmap-xxhdpi/ic_launcher_round.png
