+++ apps/mobile/src-tauri/icons/android/mipmap-xxxhdpi/ic_launcher_round.png
