+++ apps/mobile/src-tauri/icons/android/values/ic_launcher_background.xml
