+++ apps/mobile/src-tauri/src/main.rs
@@ -1,0 +1,3 @@
+fn main() {
+    vesta_mobile::run();
+}
