+++ apps/mobile/src-tauri/tauri.conf.json
@@ -1,0 +1,36 @@
+{
+  "productName": "Vesta",
+  "version": "0.1.144",
+  "identifier": "com.vestarun.mobile",
+  "build": {
+    "beforeDevCommand": "npm --workspace @vesta/web run dev",
+    "devUrl": "http://localhost:1420",
+    "beforeBuildCommand": "npm --workspace @vesta/web run build",
+    "frontendDist": "../../web/dist"
+  },
+  "app": {
+    "windows": [
+      {
+        "label": "main",
+        "title": "Vesta"
+      }
+    ],
+    "security": {
+      "csp": null
+    }
+  },
+  "bundle": {
+    "active": true,
+    "targets": "all",
+    "icon": [
+      "icons/32x32.png",
+      "icons/128x128.png",
+      "icons/128x128@2x.png",
+      "icons/icon.icns",
+      "icons/icon.ico"
+    ],
+    "iOS": {
+      "developmentTeam": "QB32XS5CT9"
+    }
+  }
+}
